refactor(output): 출력 구조 변경
변경 내용
- 서비스 계층의 출력 서비스 제거
- 컨트롤러에서 출력 뷰를 통해 값을
  출력하도록 수정
