+++ src/main/java/lotto/config/ApplicationConfiguration.java
@@ -2,7 +2,6 @@ package lotto.config;
 
 import java.util.Scanner;
 import lotto.domain.LottoMachine;
-import lotto.service.OutputService;
 import lotto.utility.RandomGenerator;
 import lotto.view.InputView;
 import lotto.view.OutputView;
@@ -14,7 +13,6 @@ public class ApplicationConfiguration {
     private final InputValidator inputValidator = new InputValidator();
     private final InputView inputView = new InputView(scanner, inputValidator);
     private final OutputView outputView = new OutputView();
-    private final OutputService outputService = new OutputService(outputView);
     private final RandomGenerator randomGenerator = new RandomGenerator();
     private final LottoMachine lottoMachine = new LottoMachine(randomGenerator);
 
@@ -22,8 +20,8 @@ public class ApplicationConfiguration {
         return inputView;
     }
 
-    public OutputService getOutputService() {
-        return outputService;
+    public OutputView getOutputView() {
+        return outputView;
     }
 
     public LottoMachine getLottoMachine() {
